requirements.txt
requests==2.23.0; python_version>="3.7"
slackclient==2.5.0; python_version>="3.7"
matplotlib==3.2.1; python_version>="3.7"
nest_asyncio==1.3.2; python_version>="3.7"